product/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.10
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.1
exceptiongroup==1.2.2
executing==2.0.1
flake8==7.1.1
idna==3.6
inflection==0.5.1
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
matplotlib-inline==0.1.7
mccabe==0.7.0
oauthlib==3.2.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.21
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
